# src/LabTracker/openapi.yaml
openapi: 3.0.0
info:
  title: Lab Management System API
  description: Полная API спецификация с учетом всех требований безопасности
  version: 2.0.0
  contact:
    name: API Support
    email: nikita2102antonenkov@gmail.com
...
paths:
...
  /courses:
    get:
      tags: [Courses]
      summary: Получить список доступных курсов
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Список курсов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Course'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Courses]
      summary: Создать новый курс (преподаватель)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseCreate'
      responses:
        '201':
          description: Курс создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /courses/{courseId}:
    get:
      tags: [Courses]
      summary: Получить информацию о курсе (участник курса)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/courseId'
      responses:
        '200':
          description: Данные курса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Доступ запрещен (не участник курса)
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Courses]
      summary: Обновить курс (преподаватель курса)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/courseId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseUpdate'
      responses:
        '200':
          description: Курс обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Доступ запрещен (не преподаватель курса)
        '404':
          $ref: '#/components/responses/NotFound'
  /courses/{courseId}/students:
    get:
      tags: [Courses]
      summary: Получить список студентов данного курса
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/courseId'
      responses:
        '200':
          description: Список курсов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CourseStudent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Доступ запрещен (не студент курса)
        '404':
          $ref: '#/components/responses/NotFound'
  /courses/{courseId}/students/{userId}:
    patch:
      tags: [Courses]
      summary: Изменить статус студента в курсе (для преподавателя)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/courseId'
        - $ref: '#/components/parameters/userId'
      responses:
        '200':
          description: Студент курса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseStudent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Доступ запрещен (не студент курса)
        '404':
          $ref: '#/components/responses/NotFound'

  # Лабораторные работы
  /courses/{courseId}/labs:
    get:
      tags: [Labs]
      summary: Получить список лабораторных работ (участник курса)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/courseId'
      responses:
        '200':
          description: Список работ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Lab'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Доступ запрещен (не участник курса)
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags: [Labs]
      summary: Создать лабораторную работу (преподаватель курса)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/courseId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LabCreate'
      responses:
        '201':
          description: Работа создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lab'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Доступ запрещен (не преподаватель курса)
        '404':
          $ref: '#/components/responses/NotFound'
  /courses/{courseId}/labs/{labId}:
    get:
      tags: [Labs]
      summary: Получить лабораторную работу (участник курса)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/courseId'
        - $ref: '#/components/parameters/labId'
      responses:
        '200':
          description: Данные работы
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lab'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Доступ запрещен (не участник курса)
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Labs]
      summary: Обновить лабораторную работу (преподаватель курса)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/courseId'
        - $ref: '#/components/parameters/labId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LabUpdate'
      responses:
        '200':
          description: Работа обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lab'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Доступ запрещен (не преподаватель курса)
        '404':
          $ref: '#/components/responses/NotFound'

  # Расписание слотов для сдачи
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Используйте JWT токен для авторизации.
        Пример: `Authorization: Bearer <token>`
  parameters:
    courseId:
      name: courseId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: ID курса
      example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
    labId:
      name: labId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: ID лабораторной работы
      example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
    userId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: ID пользователя
      example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
...
  responses:
    Unauthorized:
      description: Не авторизован
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Authorization required"
            details: ["Missing or invalid token"]
    Forbidden:
      description: Доступ запрещен
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Forbidden"
            details: ["Insufficient permissions"]
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Not found"
            details: ["Course not found"]
    BadRequest:
      description: Неверный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Bad request"
            details: ["Invalid queue mode"]
  schemas:
...
    Course:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          example: "Программная инженерия"
        description:
          type: string
          nullable: true
          example: "Основы программирования"
        teacherIds:
          type: array
          items:
            type: string
            format: uuid
          readOnly: true
        queueMode:
          type: string
          enum: [democratic, oligarchic, anarchic]
          example: "democratic"
        createdAt:
          type: string
          format: date-time
          readOnly: true
        photo:
          type: string
          format: uri
          nullable: true
          example: "https://example.com/path/to/photo.jpg"
    CourseCreate:
      type: object
      required: [name, queueMode]
      properties:
        name:
          type: string
          example: "Новый курс"
        description:
          type: string
          nullable: true
          example: "Описание курса"
        queueMode:
          type: string
          enum: [democratic, oligarchic, anarchic]
          example: "democratic"
        photo:
          type: string
          format: binary
          nullable: true
          description: "Фото курса в формате Base64"
          example: "/9j/4AAQSkZJRgABAQEAAAAAAAD/2wBDAAoHBwkICgsJEBwVGxwYHBwsJDh0dIy8sLC4qL3c3N0TCwD3AwtxgDsd0o4g73VWhu6KlgOS9JlAbxEq1eMeaGpDTKN7oPbPV0PfsEjChgT3JsbWoYG/pVREopVBAEPp7Z2FFPxl8="
    CourseUpdate:
      type: object
      properties:
        name:
          type: string
          example: "Обновленное название"
        description:
          type: string
          nullable: true
          example: "Новое описание"
        queueMode:
          type: string
          enum: [democratic, oligarchic, anarchic]
          example: "oligarchic"
        photo:
          type: string
          format: binary
          nullable: true
          description: "Фото курса в формате Base64"
          example: "/9j/4AAQSkZJRgABAQEAAAAAAAD/2wBDAAoHBwkICgsJEBwVGxwYHBwsJDh0dIy8sLC4qL3c3N0TCwD3AwtxgDsd0o4g73VWhu6KlgOS9JlAbxEq1eMeaGpDTKN7oPbPV0PfsEjChgT3JsbWoYG/pVREopVBAEPp7Z2FFPxl8="
      minProperties: 1
      additionalProperties: false
    CourseStudent:
      type: object
      properties:
        studentId:
          type: string
          format: uuid
          description: "Уникальный идентификатор студента"
        score:
          type: integer
          description: "Балл студента"
        enrollmentDate:
          type: string
          format: date-time
          description: "Дата регистрации студента на курс"
        status:
          type: string
          enum: [in_progress, completed, failed]
          description: "Статус студента в курсе"

    # Лабораторные работы
    Lab:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          example: "Лаб 1"
        description:
          type: string
          nullable: true
          example: "Основы работы с API"
        deadline:
          type: string
          format: date-time
          example: "2023-12-31T23:59:59Z"
        score:
          type: integer
          minimum: 0
          default: 100
          example: 100
        scoreAfterDeadline:
          type: integer
          minimum: 0
          default: 100
          example: 100
        courseId:
          type: string
          format: uuid
          readOnly: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
    LabCreate:
      type: object
      required: [name, deadline]
      properties:
        name:
          type: string
          example: "Новая лабораторная"
        description:
          type: string
          nullable: true
          example: "Описание работы"
        deadline:
          type: string
          format: date-time
          example: "2023-12-31T23:59:59Z"
        score:
          type: integer
          minimum: 0
          default: 100
          example: 100
        scoreAfterDeadline:
          type: integer
          minimum: 0
          default: 100
          example: 100
    LabUpdate:
      type: object
      properties:
        name:
          type: string
          example: "Обновленное название"
        description:
          type: string
          nullable: true
          example: "Новое описание"
        deadline:
          type: string
          format: date-time
          example: "2023-12-31T23:59:59Z"
        score:
          type: integer
          minimum: 0
          example: 100
        scoreAfterDeadline:
          type: integer
          minimum: 0
          example: 100
      minProperties: 1
      additionalProperties: false

    # Слот для записи
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Validation error"
        details:
          type: array
          items:
            type: string
          example: ["Email is invalid", "Password too short"]